requirements.txt
Flask
gtts
requests
python-dotenv
transformers
google-cloud-translate